+++ src/amun/gpu/decoder/encoder_decoder.h
@@ -59,7 +59,7 @@ class EncoderDecoder : public Scorer {
     const Weights& model_;
     std::unique_ptr<Encoder> encoder_;
     std::unique_ptr<Decoder> decoder_;
-    DeviceVector<size_t> indices_;
+    DeviceVector<uint> indices_;
     DeviceVector<int> batchMapping_;
       // set in Encoder::GetContext() to length (maxSentenceLength * batchSize). 1 if it's a word, 0 otherwise
